Movido ProductCard para /components, desabilitado exclusão de livros.
